Use some index.ts.  Sort BUILD rules.  Factor out UiPluginMetadata.
