+++ tensorboard/webapp/BUILD
@@ -7,97 +7,9 @@ load("//tensorboard/defs:vulcanize.bzl", "tensorboard_html_binary")
 
 licenses(["notice"])  # Apache 2.0
 
-# The app component
-tf_js_binary(
-    name = "tb_webapp_binary",
-    compile = 1,
-    entry_point = select({
-        "//tensorboard:dev_build": ":main.ts",
-        "//conditions:default": "main.prod.ts",
-    }),
-    deps = [
-        ":ng_main",
-        "//tensorboard/webapp/angular:expect_angular_material_tabs",
-        "//tensorboard/webapp/angular:expect_angular_material_toolbar",
-        "@npm//@angular/common",
-        "@npm//@angular/core",
-        "@npm//@angular/material",
-        "@npm//@angular/platform-browser",
-        "@npm//@ngrx/store",
-        "@npm//rxjs",
-        "@npm//zone.js",
-    ],
-)
 
-# the app component as a library
-tf_web_library(
-    name = "tb_webapp",
-    srcs = [
-        ":tb_webapp_binary.js",
-    ],
-    path = "/tb-webapp",
-    deps = [
-        ":tb_webapp_binary",
-    ],
-)
-
-# the complete app, including the polymer stuff
-tf_web_library(
-    name = "tensorboard-webapp",
-    srcs = [
-        "index.html",
-    ],
-    path = "/",
-    deps = [
-        "//tensorboard/components:ng_polymer_lib_binary",
-        ":tb_webapp",
-        "//tensorboard/webapp:material_theme",
-    ],
-)
-
-# html binary for the complete app
-tensorboard_html_binary(
-    name = "ng_index",
-    input_path = "/index.html",
-    output_path = "/ng_index.html",
-    deps = [":tensorboard-webapp"],
-)
-
-# TODO(stephanwlee): move this to root level
-tf_web_library(
-    name = "material_theme",
-    srcs = [
-        "@npm//:node_modules/@angular/material/prebuilt-themes/indigo-pink.css",
-    ],
-    path = "/@angular",
-    strip_prefix = "node_modules/@angular/material/prebuilt-themes",
-)
-
-
-tf_ts_library(
-    name = "config",
-    srcs = [
-        "config.dev.ts",
-        "config.prod.ts",
-    ],
-)
-
-ng_module(
-    name = "ng_main",
-    srcs = ["bootstrap.ts"] + select({
-        "//tensorboard:dev_build": ["main.ts"],
-        "//conditions:default": ["main.prod.ts"],
-    }),
-    deps = [
-        ":app",
-        ":config",
-        "@npm//@angular/core",
-        "@npm//@angular/platform-browser",
-        "@npm//@angular/router",
-        "@npm//zone.js",
-    ],
-)
-
+# Angular module for the top-level app component together with all of its
+# injectable dependencies.  I.e., the entire Angular app.
 ng_module(
     name = "app",
     srcs = [
@@ -121,6 +33,97 @@ ng_module(
         "@npm//@ngrx/effects",
         "@npm//@ngrx/store",
     ],
+)
+
+tf_ts_library(
+    name = "config",
+    srcs = [
+        "config.dev.ts",
+        "config.prod.ts",
+    ],
+)
+
+# Wrapper that prepares Angular app for deployment, dealing with browser
+# compatibility, Vulcanization, and configuration (e.g. prod vs. dev).
+ng_module(
+    name = "ng_main",
+    srcs = ["bootstrap.ts"] + select({
+        "//tensorboard:dev_build": ["main.ts"],
+        "//conditions:default": ["main.prod.ts"],
+    }),
+    deps = [
+        ":app",
+        ":config",
+        "@npm//@angular/core",
+        "@npm//@angular/platform-browser",
+        "@npm//@angular/router",
+        "@npm//zone.js",
+    ],
+)
+
+# Compile the prepared Angular app to a JS binary.
+tf_js_binary(
+    name = "tb_webapp_binary",
+    compile = 1,
+    entry_point = select({
+        "//tensorboard:dev_build": ":main.ts",
+        "//conditions:default": "main.prod.ts",
+    }),
+    deps = [
+        ":ng_main",
+        "//tensorboard/webapp/angular:expect_angular_material_tabs",
+        "//tensorboard/webapp/angular:expect_angular_material_toolbar",
+        "@npm//@angular/common",
+        "@npm//@angular/core",
+        "@npm//@angular/material",
+        "@npm//@angular/platform-browser",
+        "@npm//@ngrx/store",
+        "@npm//rxjs",
+        "@npm//zone.js",
+    ],
+)
+
+# Wrap the Angular app JS binary as a library.
+tf_web_library(
+    name = "tb_webapp",
+    srcs = [
+        ":tb_webapp_binary.js",
+    ],
+    path = "/tb-webapp",
+    deps = [
+        ":tb_webapp_binary",
+    ],
+)
+
+tf_web_library(
+    name = "material_theme",
+    srcs = [
+        "@npm//:node_modules/@angular/material/prebuilt-themes/indigo-pink.css",
+    ],
+    path = "/@angular",
+    strip_prefix = "node_modules/@angular/material/prebuilt-themes",
+)
+
+# Bundle the Angular app with the Polymer components and plugins, as a library.
+tf_web_library(
+    name = "tensorboard-webapp",
+    srcs = [
+        "index.html",
+    ],
+    path = "/",
+    deps = [
+        "//tensorboard/components:ng_polymer_lib_binary",
+        ":tb_webapp",
+        "//tensorboard/webapp:material_theme",
+    ],
+)
+
+# A Vulcanized html binary for the complete app (both Angular and Polymer parts)
+tensorboard_html_binary(
+    name = "ng_index",
+    input_path = "/index.html",
+    output_path = "/ng_index.html",
+    deps = [":tensorboard-webapp"],
 )
 
 # Karma has overhead of bootstrap/tearDown. Combine as much testcases
