+++ tensorboard/webapp/core/effects/core.effects.ts
@@ -1,8 +1,11 @@
 /* Copyright 2019 The TensorFlow Authors. All Rights Reserved.
+
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
+
     http://www.apache.org/licenses/LICENSE-2.0
+
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -28,8 +31,8 @@ import {
   pluginsListingRequested,
   pluginsListingLoaded,
   pluginsListingFailed,
-} from '../actions/core.actions';
-import {State, getPluginsListLoaded} from '../store/core.reducers';
+} from '../actions';
+import {State, getPluginsListLoaded} from '../store';
 import {LoadState} from '../../types/api';
 
 /** @typehack */ import * as _typeHackRxjs from 'rxjs';
