+++ tensorboard/webapp/header/types.ts
@@ -13,4 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-export * from './core.reducers';
+import {PluginMetadata, PluginId} from '../types/api';
+
+export interface UiPluginMetadata extends PluginMetadata {
+  id: PluginId;
+}
