+++ tensorboard/webapp/plugins/plugins.container.test.ts
@@ -20,7 +20,7 @@ import {provideMockStore, MockStore} from '@ngrx/store/testing';
 import {PluginsContainer} from './plugins.container';
 
 import {PluginId, LoadingMechanismType, LoadState} from '../types/api';
-import {State, CoreState, CORE_FEATURE_KEY} from '../core/store/core.reducers';
+import {State, CoreState, CORE_FEATURE_KEY} from '../core/store';
 import {createState, createCoreState} from '../core/testing';
 
 /** @typehack */ import * as _typeHackStore from '@ngrx/store';
