+++ tensorboard/webapp/settings/dialog.component.ts
@@ -19,15 +19,8 @@ import {Store, select, createSelector} from '@ngrx/store';
 import {Subject} from 'rxjs';
 import {takeUntil, debounceTime, filter} from 'rxjs/operators';
 
-import {
-  State,
-  getReloadEnabled,
-  getReloadPeriodInMs,
-} from '../core/store/core.reducers';
-import {
-  toggleReloadEnabled,
-  changeReloadPeriod,
-} from '../core/actions/core.actions';
+import {State, getReloadEnabled, getReloadPeriodInMs} from '../core/store';
+import {toggleReloadEnabled, changeReloadPeriod} from '../core/actions';
 
 /** @typehack */ import * as _typeHackRxjs from 'rxjs';
